Clean out kwargs dict in cloud.action before calling cloud driver function
Fixes #40278

When we run a cloud.action function from the CLI, the various __pub_* keys
and values populate the "kwargs" arg. Then, when we attempt to call out
to a cloud driver's function that doesn't accept a "kwarg" arg (or there are
too many args passed now), we get an error.

If the cloud function only takes "name" and "call", we should not be passing
in "kwargs", too.